+++ src/content-scripts/page-adapters/chatgpt.ts
@@ -1,11 +1,32 @@
 import PageLive from "../pagelive";
+import ElementObserver from "../element-observer";
 
 const chatgptAdapter = async () => {
     const pl = new PageLive();
-    
+
     const construct = async () => {
-        console.log("ChatGPT page adapter is active!");
+
+        observeForShortcutModal();
+
+        await pl.page.ready();
     }
+    const observeForShortcutModal = () => {
+        const shortcutModal = new ElementObserver(
+            async () => document.querySelector('[data-testid="keyboard-shortcuts-dialog"]'),
+            // callback when dialog is opened.
+            async (element) => {
+                // No need to announce to users since SR will automatically announce the dialog when it is opened. In fact, announcing here will cause duplicate announcements which can be annoying for users.
+
+                // Modify the shortcut to be accessible for SR users.
+            },
+            null,
+            // Callback when dialog is closed.
+            async (element) => {
+                pl.speak("Keyboard shortcuts dialog is closed");
+            },
+        );
+        shortcutModal.observe();
+    };
     await construct();
 }
 
@@ -15,3 +36,5 @@ document.addEventListener('DOMContentLoaded', () => {
         console.error("Error initializing ChatGPT page adapter:", err);
     });
 });
+
+